Remove vtarget.dto module - only had one class, moved up
