--- api/src/main/java/org/jmisb/api/klv/st0903/vtarget/dto/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Data Transfer Objects for ST 0903 VTarget Local Set.
- */
-package org.jmisb.api.klv.st0903.vtarget.dto;
